feat: create order methods in a separate file
