+++ products/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
 from . import views
 from . import carts_views
+from . import cart_create_order
 
 urlpatterns = [
     path ('', views.home, name='home'),
@@ -14,7 +15,7 @@ urlpatterns = [
     path('cart/update/', carts_views.update_cart, name='update_cart'),
     path('cart/get-items/', carts_views.get_cart_items, name='get_cart_items'),
     path('cart/remove/', carts_views.remove_cart_item, name='remove_cart_item'),
-    path('cart/order/', carts_views.create_order, name='create_order'),
+    path('cart/order/', cart_create_order.create_order, name='create_order'),
     path('cart/determine-contribution-usage/', carts_views.determine_contribution_usage, name='determine_contribution_usage'),
     path('orders/', views.orders, name='orders'),
     path('orders/update-status/', views.update_order_status, name='update_order_status'),
